:tractor: relocate test case
